Simplify and change parsing (#25)
Parsing the ingredients and procedure has been pretty error prone in the
past. Historical PlateZero really tried to take the amount given and put
it into a fraction, as the only well known way to represent the amount.

Knowing the amount of ingredient and the unit allows for multiplying of
the ingredients (double the recipe) and also converting between units.

But it only ever did the first number. Which meant if you had `1 lb` of
asparagus, and that was about ~12 spears, it might double the first
amount but not touch the second.

It is fragile and error prone.

The new short term solution is to simply remove all this. Ingredients
and instructions have three different values:

1. `raw` - The exact thing we parse with no changes. By keeping this
   forever, we can recalculate the next two as the algorithm changes.
   Very helpful.

2. `normalized` - A cleaned up version of `raw`. This will be the next
   thing worked on and will probably keep things good for a long time.
   Trim strings, replace HTML entities, remove HTML, cleanup
   punctutation. Very basic things. But what you would expect to see
   after parsing. Importantly, this will be calculated from raw so if we
   change how we do it then we can run it for all recipes again.

3. `enhanced` - A future feature where the code will go back through and
   parse out those entities for enhanced features. Understand numbers,
   units, optional ingredients, and more. Do this in a smarter way
   though, and always calculate it from `normalized`.